labwork8/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.5.1
pymongo==4.8.0
pydantic==2.5.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis[asyncio]==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2